+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Allow flatcar alpha releases for machine pools.
+- Allow newer flatcar releases for node pools as provided by AWS release.
 
 ## [14.19.2] - 2023-08-03